ha/requirements.txt
# duck duck go search
ddgs

# Voice - stt
vosk
sounddevice

# Voice tts
piper-tts

# ADK
google-adk
#   ADK optional dependency
litellm

